Moved implementation to separate graphviz.cpp file
